test(cost): split calculator tests for GH727
